+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setup tests with node:test runner. Usage: `yarn test` and `yarn test:watch`.
 - Add github workflows.
 
+### Changed
+
+- Switch from cookie to authorization header in `v7.createAccount`.
+
 ## [0.1.1] - 2025-03-26
 
 ### Fixed